Move media validation to contracts
