+++ backend/app/infrastructure/transcription/audio_processing.py
@@ -3,25 +3,18 @@ Audio extraction and processing utilities
 """
 
 import asyncio
-import json
 import logging
 import os
-import platform
 import subprocess
 import tempfile
-from types import ModuleType
 
 from django.conf import settings
-from django.core.exceptions import ImproperlyConfigured
 from openai import AsyncOpenAI
-
-resource_module: ModuleType | None
-try:
-    import resource
-except ImportError:  # pragma: no cover - non-Unix fallback
-    resource_module = None
-else:
-    resource_module = resource
+from app.contracts.media_validation import (
+    InvalidMediaFileError,
+    probe_media_file,
+    run_media_command,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -39,166 +32,20 @@ SUPPORTED_FORMATS = {
     ".webm",
 }
 
-ALLOWED_VIDEO_CONTAINER_FORMATS = {
-    "3gp",
-    "3g2",
-    "avi",
-    "matroska",
-    "mov",
-    "mp4",
-    "mpeg",
-    "webm",
-}
-
-ALLOWED_VIDEO_CODECS = {
-    "av1",
-    "h263",
-    "h264",
-    "hevc",
-    "mpeg4",
-    "msmpeg4v3",
-    "vp8",
-    "vp9",
-}
-
-
-class InvalidMediaFileError(ValueError):
-    """Raised when an uploaded file cannot be verified as a supported video."""
-
-
-def _get_setting(name, default):
-    try:
-        return getattr(settings, name, default)
-    except ImproperlyConfigured:
-        return default
-
-
-def _build_media_preexec_fn():
-    """
-    Apply conservative resource limits to ffmpeg/ffprobe child processes on Unix.
-    """
-    if platform.system() == "Windows" or resource_module is None:
-        return None
-
-    cpu_time_limit_seconds = int(
-        _get_setting("MEDIA_PROCESS_CPU_TIME_LIMIT_SECONDS", 30)
-    )
-    memory_limit_mb = int(_get_setting("MEDIA_PROCESS_MEMORY_LIMIT_MB", 1024))
-    output_file_size_limit_mb = int(
-        _get_setting("MEDIA_PROCESS_OUTPUT_FILE_SIZE_LIMIT_MB", 512)
-    )
-
-    def _set_limits():
-        if cpu_time_limit_seconds > 0:
-            resource_module.setrlimit(
-                resource_module.RLIMIT_CPU,
-                (cpu_time_limit_seconds, cpu_time_limit_seconds),
-            )
-        if memory_limit_mb > 0:
-            memory_limit_bytes = memory_limit_mb * 1024 * 1024
-            resource_module.setrlimit(
-                resource_module.RLIMIT_AS,
-                (memory_limit_bytes, memory_limit_bytes),
-            )
-        if output_file_size_limit_mb > 0:
-            output_file_limit_bytes = output_file_size_limit_mb * 1024 * 1024
-            resource_module.setrlimit(
-                resource_module.RLIMIT_FSIZE,
-                (output_file_limit_bytes, output_file_limit_bytes),
-            )
-
-    return _set_limits
-
-
-def _run_media_command(command, timeout_seconds):
-    return subprocess.run(
-        command,
-        capture_output=True,
-        text=True,
-        check=True,
-        timeout=timeout_seconds,
-        preexec_fn=_build_media_preexec_fn(),
-    )
-
-
-def probe_media_file(input_path, timeout_seconds=10):
-    """
-    Inspect a media file with ffprobe and return parsed metadata.
-    """
-    try:
-        probe_result = _run_media_command(
-            [
-                "ffprobe",
-                "-v",
-                "quiet",
-                "-print_format",
-                "json",
-                "-show_format",
-                "-show_streams",
-                input_path,
-            ],
-            timeout_seconds,
-        )
-    except (subprocess.CalledProcessError, subprocess.TimeoutExpired, OSError) as exc:
-        raise InvalidMediaFileError("Uploaded file could not be validated as a video.") from exc
-
-    try:
-        return json.loads(probe_result.stdout)
-    except json.JSONDecodeError as exc:
-        raise InvalidMediaFileError("Uploaded file could not be validated as a video.") from exc
-
-
-def validate_video_media_file(
-    input_path,
-    allowed_container_formats=None,
-    timeout_seconds=10,
-):
-    """
-    Validate that a file is a supported video container with at least one video stream.
-    """
-    probe = probe_media_file(input_path, timeout_seconds=timeout_seconds)
-    streams = probe.get("streams") or []
-    format_info = probe.get("format") or {}
-
-    video_streams = [
-        stream for stream in streams if stream.get("codec_type") == "video"
-    ]
-    if not video_streams:
-        raise InvalidMediaFileError("Uploaded file is not a valid video.")
-
-    codec_names = {stream.get("codec_name") for stream in video_streams}
-    if not all(codec_names):
-        raise InvalidMediaFileError("Uploaded file is missing video codec metadata.")
-    if not codec_names.issubset(ALLOWED_VIDEO_CODECS):
-        raise InvalidMediaFileError("Uploaded video codec is not supported.")
-
-    format_names = {
-        name.strip()
-        for name in (format_info.get("format_name") or "").split(",")
-        if name.strip()
-    }
-    allowed_formats = set(allowed_container_formats or ALLOWED_VIDEO_CONTAINER_FORMATS)
-    if not format_names or format_names.isdisjoint(allowed_formats):
-        raise InvalidMediaFileError("Uploaded video container is not supported.")
-
-    duration_raw = format_info.get("duration")
-    try:
-        duration = float(duration_raw)
-    except (TypeError, ValueError) as exc:
-        raise InvalidMediaFileError("Uploaded file is missing a valid duration.") from exc
-
-    if duration <= 0:
-        raise InvalidMediaFileError("Uploaded file duration must be greater than zero.")
-
-    return probe
-
-
 def _get_video_duration(input_path):
     """
     Get video duration using ffprobe
     Returns: duration in seconds
     """
-    probe = probe_media_file(input_path)
+    probe = probe_media_file(
+        input_path,
+        timeout_seconds=int(getattr(settings, "FFPROBE_VALIDATION_TIMEOUT_SECONDS", 10)),
+        cpu_time_limit_seconds=int(getattr(settings, "MEDIA_PROCESS_CPU_TIME_LIMIT_SECONDS", 30)),
+        memory_limit_mb=int(getattr(settings, "MEDIA_PROCESS_MEMORY_LIMIT_MB", 1024)),
+        output_file_size_limit_mb=int(
+            getattr(settings, "MEDIA_PROCESS_OUTPUT_FILE_SIZE_LIMIT_MB", 512)
+        ),
+    )
     duration = float(probe["format"]["duration"])
     logger.info(f"Video duration: {duration:.2f} seconds")
     return duration
@@ -222,7 +69,7 @@ def _extract_full_audio(input_path, temp_dir):
         temp_dir, f"temp_audio_{os.path.basename(input_path)}.mp3"
     )
 
-    _run_media_command(
+    run_media_command(
         [
             "ffmpeg",
             "-i",
@@ -235,7 +82,12 @@ def _extract_full_audio(input_path, temp_dir):
             "-y",
             temp_audio_path,
         ],
-        int(_get_setting("FFMPEG_PROCESS_TIMEOUT_SECONDS", 120)),
+        timeout_seconds=int(getattr(settings, "FFMPEG_PROCESS_TIMEOUT_SECONDS", 120)),
+        cpu_time_limit_seconds=int(getattr(settings, "MEDIA_PROCESS_CPU_TIME_LIMIT_SECONDS", 30)),
+        memory_limit_mb=int(getattr(settings, "MEDIA_PROCESS_MEMORY_LIMIT_MB", 1024)),
+        output_file_size_limit_mb=int(
+            getattr(settings, "MEDIA_PROCESS_OUTPUT_FILE_SIZE_LIMIT_MB", 512)
+        ),
     )
 
     audio_size_mb = os.path.getsize(temp_audio_path) / (1024 * 1024)
@@ -268,7 +120,7 @@ def _extract_audio_segment(input_path, start_time, end_time, segment_index, temp
         temp_dir, f"audio_segment_{segment_index}_{os.path.basename(input_path)}.mp3"
     )
 
-    _run_media_command(
+    run_media_command(
         [
             "ffmpeg",
             "-ss",
@@ -285,7 +137,12 @@ def _extract_audio_segment(input_path, start_time, end_time, segment_index, temp
             "-y",
             audio_path,
         ],
-        int(_get_setting("FFMPEG_PROCESS_TIMEOUT_SECONDS", 120)),
+        timeout_seconds=int(getattr(settings, "FFMPEG_PROCESS_TIMEOUT_SECONDS", 120)),
+        cpu_time_limit_seconds=int(getattr(settings, "MEDIA_PROCESS_CPU_TIME_LIMIT_SECONDS", 30)),
+        memory_limit_mb=int(getattr(settings, "MEDIA_PROCESS_MEMORY_LIMIT_MB", 1024)),
+        output_file_size_limit_mb=int(
+            getattr(settings, "MEDIA_PROCESS_OUTPUT_FILE_SIZE_LIMIT_MB", 512)
+        ),
     )
 
     return {"path": audio_path, "start_time": start_time, "end_time": end_time}
